+++ modules/core/src/service/dynamic-form.service.ts
@@ -1,8 +1,8 @@
 import {Injectable} from "@angular/core";
 import {FormBuilder, FormControl, FormGroup, FormArray, Validators} from "@angular/forms";
-import {IDynamicFormGroupModel} from "../model/dynamic-form.model";
+import {DynamicFormAbstractControlModel} from "../model/dynamic-form-abstract-control.model";
 import {DynamicFormControlModel} from "../model/dynamic-form-control.model";
-import {DynamicFormArrayModel, DYNAMIC_FORM_CONTROL_TYPE_ARRAY} from "../model/dynamic-form-array.model";
+import {DynamicFormArrayModel, DYNAMIC_FORM_CONTROL_TYPE_ARRAY} from "../model/array/dynamic-form-array.model";
 import {
     DYNAMIC_FORM_CONTROL_TYPE_CHECKBOX_GROUP,
     DynamicCheckboxGroupModel
@@ -18,22 +18,27 @@ export class DynamicFormService {
         this.formBuilder = formBuilder;
     }
 
+    createFormArrayGroup (dynamicFormArrayModel: DynamicFormArrayModel): FormGroup {
+
+        return this.createFormGroup(dynamicFormArrayModel.addGroup());
+    }
+
     createFormArray(dynamicFormArrayModel: DynamicFormArrayModel, count: number): FormArray {
 
         let formArray = [];
 
         for (let i = 0; i < count; i += 1) {
-            formArray.push(this.createFormGroup(dynamicFormArrayModel));
+            formArray.push(this.createFormArrayGroup(dynamicFormArrayModel));
         }
 
         return this.formBuilder.array(formArray);
     }
 
-    createFormGroup(formGroupModel: IDynamicFormGroupModel): FormGroup {
+    createFormGroup(group: Array<DynamicFormAbstractControlModel>): FormGroup {
 
         let formGroup = {};
 
-        formGroupModel.group.forEach(model => {
+        group.forEach(model => {
 
             if (model.type === DYNAMIC_FORM_CONTROL_TYPE_ARRAY) {
 
@@ -43,7 +48,9 @@ export class DynamicFormService {
 
             } else if (model.type === DYNAMIC_FORM_CONTROL_TYPE_CHECKBOX_GROUP) {
 
-                formGroup[model.id] = this.createFormGroup(<DynamicCheckboxGroupModel> model);
+                let checkboxGroupModel = <DynamicCheckboxGroupModel> model;
+
+                formGroup[model.id] = this.createFormGroup(checkboxGroupModel.group);
 
             } else {